requirements.txt
fastapi
uvicorn
httpx
python-multipart
pydantic
databases
asyncpg
requests
sqlmodel
passlib[bcrypt]
google-auth
python-dotenv
toml
websockets==10.4
psycopg2-binary
pgvector
google-generativeai
redis
